Move FileFetcherJob to datastore
